routing through post slugs + markdown HTML transformer
