+++ package.json
@@ -34,6 +34,7 @@
     "gatsby-plugin-styled-components": "^5.19.0",
     "gatsby-source-filesystem": "^3.14.0",
     "masonry-layout": "^4.2.2",
+    "micromark": "^3.1.0",
     "react": "^17.0.1",
     "react-chartjs-2": "^4.3.1",
     "react-dom": "^17.0.1",
