+++ src/components/blog-cards/style.scss
@@ -18,6 +18,7 @@
         background-color: transparent;
         background-repeat: no-repeat;
         background-size: cover;
+        background-position: center;
     }
 }
