+++ src/services/StoryblockService.js
@@ -1,0 +1,16 @@
+import StoryblokClient from "storyblok-js-client";
+
+const StoryblockService = () => {
+
+    if (typeof StoryblokInstance !== "object") {
+        const StoryblokInstance = new StoryblokClient({
+            accessToken: process.env.STORYBLOK_TOKEN,
+            region: process.env.STORYBLOK_REGION
+        });
+        return StoryblokInstance;
+    } else {
+        return StoryblokInstance;
+    }
+}
+
+export default StoryblockService;
