+++ src/utils/Micromark.js
@@ -1,0 +1,5 @@
+import {micromark} from 'micromark'
+
+export const Micromark = ( mdContent ) => {
+    return micromark(mdContent)
+}
